Merge tests together so they're not run in parallel
Lab doesn't seem to support running tests one after another
sequentially.  This means that the auth test and the end-to-end post
test will always stomp over one another's databases and such.

I'm sure there's probably a better lab-ish way to do this, but this gets
the tests to stop messing with one another.