Move config loading to ./lib/conf.js
If the `nconf.argv().env().file({ file: 'local.json' })` dance is done
in multiple places, then each place wipes out the configs loaded by the
last run.

So, if you want to override the port or db root or other fun stuff in
testing, it's super tricky.

With this change, a test can do `require('../lib/conf').set(..)` and
it'll Just Work.

Granted, most configs are only actually *used* once at startup, so it
has to be the first thing that the test does, prior to loading any
modules or routes.