Reply deletion frontend controls (not functional)
Put a little red X next to reply links when you are the owner of the
post, so that you can delete the replies if they bother you.

This commit is only the front-end portions.  Also move the reply html
generation to views/post.jade, where it belongs.

Todo: actually delete reply link.