Simplify Stylus
This uses npm scripts rather than grunt to do the stylus build.  Removes quite a few dependencies and scripts to maintain.

This is not yet ready to merge as it is tracking off the master of stylus.  We will need to wait for them to cut a new version before we should land this.