Add tab-completion in the chat input (closes #76)
This patch adds a nice little tab-completion to BBS chat. If the tab key
is pushed while typing a chat message, instead of switching to the
address bar, it'll search through the users list and complete a name if
the name is found.

As a quick note... This currently only completes the last word as a
name... I haven't tried doing anything around completing based on cursor
position, but I could spend more time on this if that's important. My
guess is that this should be okay for a basic feature, as usually tab
completion is attempted on the fly!

Signed-off-by: David Celis <me@davidcel.is>